feat(web): vault「复制接入包」升级为完整包——与「让 agent 加入」共用同一 builder
#584 的复制现场重建当时用了最小包，缺 charter 快照与待命/唤醒指引：从 vault 复制
接入的新 agent 报到完就不知道怎么挂 watch/serve。现把完整包构造抽成
lib/joinPack.buildFullJoinPack，两个入口调同一份，产物逐字节同构：

- joinPack.ts 独立成模块（AgentJoin 测试整体 mock 了 vault，builder 放那边会被假实现
  顶掉）；MIN_CLI / VERSION_GE_SNIPPET 真值随迁，vault re-export 保住既有消费者
- AgentTokens 新增 charter prop（Channel 页与 AgentJoin 同源透传，复制零异步零失败面）
- copypack 测试升级：断言复制产物含 charter 正文、watch --once 待命行、party_* 指引
