+++ web/src/components/AgentTokens.copypack.test.tsx
@@ -132,7 +132,7 @@ async function renderOpen(): Promise<ReactTestRenderer> {
   await act(async () => {
     r = create(
       <LocaleProvider>
-        <AgentTokens slug="demo" token="tok-1" accountKey="acct-1" inviterName="host" onAuthFailed={() => {}} />
+        <AgentTokens slug="demo" token="tok-1" accountKey="acct-1" inviterName="host" charter={{ charter: "read the pinned rules before posting", rev: 3 } as never} onAuthFailed={() => {}} />
       </LocaleProvider>,
       {
         createNodeMock(element) {
@@ -176,6 +176,12 @@ describe("AgentTokens copy join pack (#584)", () => {
     expect(pack).toContain("claude mcp add party-legacy-bot --env");
     expect(pack).toContain("--token ap_old_token");
     expect(pack).toContain("party init --server https://party.example");
+    // ……而且是与「＋ 让 agent 加入」同构的【完整包】：charter 快照 + 待命/唤醒指引 + 参与指引，
+    // 不是只有 init/check-in 的最小包（否则新 agent 报到完就不知道怎么挂 watch/serve）。
+    expect(pack).toContain("read the pinned rules before posting");
+    expect(pack).toContain("party watch demo --mentions-only --once");
+    expect(pack).toContain("party_decision_ask");
+    expect(pack).toContain('party send "');
     // ……而不是 vault 里的冻结文本（TMPDIR 路径 / 0.2.52 是旧包指纹）。
     expect(pack).not.toContain("TMPDIR");
     expect(pack).not.toContain("0.2.52");
